+++ Android/SingMate/app/src/main/java/com/sigorzav/singmate/data/remote/song/SongAPI.kt
@@ -1,8 +1,9 @@
 package com.sigorzav.singmate.data.remote.song
 
-import com.sigorzav.singmate.model.ApiResponse
+import com.sigorzav.singmate.model.response.ApiResponse
 import com.sigorzav.singmate.model.Song
 import retrofit2.http.GET
+import retrofit2.http.POST
 import retrofit2.http.Query
 
 interface SongAPI {
@@ -10,6 +11,12 @@ interface SongAPI {
     /**
      * 노래 검색 (목록 조회)
      */
-    @GET("song/search")
-    suspend fun searchSongs(@Query("query") query: String): ApiResponse<Song>
+    @GET("api/v1/song/search")
+    suspend fun searchSongs(@Query("query") query: String): ApiResponse<List<Song>>
+
+    /**
+     * 노래 상세 조회
+     */
+    @POST("song/songDetail")
+    suspend fun songDetail(@Query("song") song: Song): Song
 }
